lab/docker/api-lab/requirements.txt
flask==3.0.0
flask-graphql==2.0.1
graphene==3.3
psycopg2-binary==2.9.9
pyjwt==2.8.0
flask-cors==4.0.0
